First real attempt at proper locking.  The locking is a little complicated since the the command and data that is being built to be sent to or read from the HW lives in the softc.  Commands are later run via an_setdef etc. In the ioctl path various references are kept to the data stored in the softc so it needs to be protected.  Almost think of the command in the softc a global variable since it essentially is.  Since locking wasn't done in this type of context the commands would get corrupted.
Thanks to avatar@ for catching some lock issues and dhw@ for testing.

Things are a lot more stable except for the MPI-350 cards.  My an(4)
remote laptop stays on the network now.

The driver should be changed so that it uses private memory that is passed
to the functions that talk to the card.  Then only those functions would
really need to grab locks.

Reviewed by:	avatar@


git-svn-id: svn://svn.freebsd.org/base/head@175445 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f